- Added NBT - Added b1.7.3 world generation - Added b1.7.3 McRegion save format - Added a lot of blocks from b1.7.3 (still incomplete, mainly the behavior/ rendering) - Adde almost all items from b1.7.3 (few missing or without function/inaccurate) - Added Container Menus and now the Inventory is accurate - Added Language system (using json instead properties) - Added basic Sound loading system - Added Tile Entities - Added Tile Entities Renderers - Added crafting (missing a lot of recipes) - Added furnace recipes - Fixed game not pausing, even if theoretically offline (the game will pause whenever only one player is online)

+++ source/world/level/storage/LevelStorageSource.hpp
@@ -14,8 +14,8 @@ struct LevelSummary
 {
 	std::string m_fileName;
 	std::string m_levelName;
-	int m_lastPlayed;
-	int m_sizeOnDisk;
+	int64_t m_lastPlayed;
+	int64_t m_sizeOnDisk;
 
 	LevelSummary()
 	{
@@ -23,7 +23,7 @@ struct LevelSummary
 		m_sizeOnDisk = 0;
 	}
 
-	LevelSummary(const std::string& a, const std::string& b, int c, int d)
+	LevelSummary(const std::string& a, const std::string& b, int64_t c, int64_t d)
 	{
 		m_fileName = a;
 		m_levelName = b;
